AMAZONscrapy/requirements.txt
requests==2.9.1
scrapy-redis==0.6.8
redis==2.10.6
lxml==4.2.5
pymongo==3.7.1
PyMySQL==0.9.2




